Control.Monad.Trace

Documentation

type TraceMaybe a = MaybeT (Writer (Last a))

Trace is a neat way to run a Maybe monad, but leave a trail behind so that if it fails, we know what step it failed at.

withTrace :: Monad m => Maybe a -> MaybeT m a

attempt :: a -> TraceMaybe a ()

runTrace :: e -> TraceMaybe e a -> Either e a