Language.Haskell.Interpreter

Documentation

runghc :: (Show t, TimeUnit t, MonadIO m, MonadError InterpreterError m, MonadMask m) => t -> [String] -> FilePath -> m (InterpreterResult String)

spawn an external process to runghc a file

If you set the environmental varaiable GHC_BIN_DIR then the executable runghc in that path will be used. This is useful if you want to your file to be run with some packages available, you can create a wrapper runghc that includes these

Any errors are converted to InterpreterError

data CompilationError

Constructors

RawError Text

CompilationError

Fields

filename :: !Text

row :: !Int

column :: !Int

text :: ![Text]

data InterpreterError

Constructors

CompilationErrors [CompilationError]

TimeoutError Text

newtype SourceCode

Constructors

SourceCode Text

avoidUnsafe :: MonadError InterpreterError m => SourceCode -> m ()

newtype Warning

Constructors

Warning Text

data InterpreterResult a

Constructors

InterpreterResult

Fields

warnings :: [Warning]

result :: a

parseErrorText :: Text -> CompilationError