Playground.Schema

Description

This module handles exposing a Contract API to the Plutus Playground frontend.

In practice this means having a way of turning the Contract's effect rows into a Schema declaration, by using an EndpointToSchema instance.

|

Documentation

endpointsToSchemas :: EndpointToSchema s => [FunctionSchema FormSchema]

class EndpointToSchema (s :: Row Type)

Minimal complete definition

endpointsToSchemas