Playground.Types

Documentation

data KnownCurrency

Constructors

KnownCurrency

Fields

hash :: ValidatorHash

friendlyName :: String

knownTokens :: NonEmpty TokenName

adaCurrency :: KnownCurrency

data PayToWalletParams

Constructors

PayToWalletParams

Fields

payTo :: WalletNumber

value :: Value

data SimulatorWallet

Constructors

SimulatorWallet

Fields

simulatorWalletWallet :: WalletNumber

simulatorWalletBalance :: Value

data ContractCall a

Describes the mockchain requests the frontend can make of the backend. These will be mostly calls to their contract's various endpoints, but we supply a few extra special calls for the sake of easier testing and simulation.

Constructors

CallEndpoint
Call one of the defined endpoints of your contract.

Fields

caller :: WalletNumber

argumentValues :: FunctionSchema a

AddBlocks
Add the specified number of blocks to the mockchain before continuing.

Fields

blocks :: Integer

AddBlocksUntil
Keep adding blocks until the mockchain reaches the specified slot, then continue. (Note that calling AddBlocksUntil 20 doesn't mean you'll continue at slot 20, just that the slot number will now be at least that high.

Fields

slot :: Slot

PayToWallet
Make a wallet-to-wallet transfer of the specified value.

Fields

sender :: WalletNumber

recipient :: WalletNumber

amount :: Value

type SimulatorAction = ContractCall (Fix FormArgumentF)

type Expression = ContractCall Value

data Simulation

The example simulations have meaningful (hard-coded) names, but, for any simulation the user creates in the playground, simulationName = "Simulation " <> show simulationId; the simulationId is simply present to ensure every simulation is created with a unique name.

Constructors

Simulation

Fields

simulationName :: String

simulationId :: Int

simulationActions :: [SimulatorAction]

simulationWallets :: [SimulatorWallet]

data Evaluation

Constructors

Evaluation

Fields

wallets :: [SimulatorWallet]

sourceCode :: SourceCode

program :: Value

This will be a '[Expression s]' where s is the schema from the compiled SourceCode. It has to be JSON, because we can't know the type of s until the SourceCode has been compiled.

pubKeys :: Evaluation -> [PaymentPubKeyHash]

data EvaluationResult

Constructors

EvaluationResult

Fields

resultRollup :: [[AnnotatedTx]]

Annotated blockchain, newest blocks first

emulatorLog :: [EmulatorEvent]

The emulator log, newest events first

emulatorTrace :: Text

fundsDistribution :: [SimulatorWallet]

feesDistribution :: [SimulatorWallet]

walletKeys :: [(PaymentPubKeyHash, WalletNumber)]

data CompilationResult

Constructors

CompilationResult

Fields

functionSchema :: [FunctionSchema FormSchema]

knownCurrencies :: [KnownCurrency]

data ContractDemo

Constructors

ContractDemo

Fields

contractDemoName :: Text

contractDemoEditorContents :: SourceCode

contractDemoSimulations :: [Simulation]

contractDemoContext :: InterpreterResult CompilationResult

data FunctionSchema a

Constructors

FunctionSchema

Fields

endpointDescription :: EndpointDescription

argument :: a

All contract endpoints take a single argument. (Multiple arguments must be wrapped up into a container.)

data PlaygroundError

Constructors

CompilationErrors [CompilationError]

InterpreterError InterpreterError

RollupError Text

OtherError String

JsonDecodingError

Fields

expected :: String

decodingError :: String

input :: String